changed variable type

+++ A10.cpp
@@ -7,7 +7,7 @@ int main()
 {
     ios_base::sync_with_stdio(false);
     cin.tie(0);
-    int n,t;
+    long n,t;
     cin>>n>>t;
     string s;
     cin>>s;
